# packages/api-contracts/openapi.yaml
...
openapi: 3.0.0
paths:
...
  /api/v1/notifications:
    get:
      description: Requires authentication. meta.unreadCount is the total unread count regardless of the unread filter.
      operationId: NotificationController_list_v1
      parameters:
        - name: unread
          required: false
          in: query
          schema:
            type: string
          description: Pass 'true' or '1' to return only unread items. Any other value is treated as false; not schema-validated.
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NotificationResponseDto'
                  meta:
                    $ref: '#/components/schemas/NotificationMetaDto'
                required:
                  - success
                  - data
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      security:
        - bearer: []
      summary: List the signed-in user's notifications
      tags:
        - notifications
  /api/v1/notifications/read-all:
    patch:
      description: Requires authentication.
      operationId: NotificationController_readAll_v1
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/MarkedResultDto'
                required:
                  - success
                  - data
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      security:
        - bearer: []
      summary: Mark every notification as read
      tags:
        - notifications
  /api/v1/notifications/{id}/read:
    patch:
      description: Requires authentication.
      operationId: NotificationController_read_v1
      parameters:
        - name: id
          required: true
          in: path
          description: Notification id.
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/NotificationResponseDto'
                required:
                  - success
                  - data
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '404':
          description: NOT_FOUND — the requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      security:
        - bearer: []
      summary: Mark one notification as read
      tags:
        - notifications
...
  /api/v1/products/history:
    get:
      description: Public (guest). Real recorded snapshots only — returns an empty list when none exist yet (and when productId is omitted). Never interpolated.
      operationId: PriceHistoryController_history_v1
      parameters:
        - name: productId
          required: false
          in: query
          schema:
            type: string
          description: Internal product UUID. Not schema-validated; omitting it returns an empty list rather than an error.
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PricePointDto'
                  meta:
                    $ref: '#/components/schemas/CountMetaDto'
                required:
                  - success
                  - data
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Daily price snapshots for a product
      tags:
        - products
  /api/v1/products/price-graph:
    get:
      description: Public (guest). One series per seller with recorded history. Empty when productId is omitted or nothing is recorded yet.
      operationId: PriceHistoryController_priceGraph_v1
      parameters:
        - name: productId
          required: false
          in: query
          schema:
            type: string
          description: Internal product UUID. Not schema-validated; omitting it returns an empty series list.
        - name: days
          required: false
          in: query
          schema:
            type: string
          description: 'Range in days. NOT validated: any value outside 30 | 90 | 180 | 365 | 1095 is silently clamped to 90 rather than rejected. meta.days reports what was actually applied.'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/PriceGraphResponseDto'
                  meta:
                    $ref: '#/components/schemas/PriceGraphMetaDto'
                required:
                  - success
                  - data
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Per-seller price history series
      tags:
        - products
...
info:
  title: Omnisift API
  description: Product Intelligence Platform — REST API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - AUTHENTICATION_ERROR
        - AUTHORIZATION_ERROR
        - NOT_FOUND
        - PROVIDER_ERROR
        - DATABASE_ERROR
        - BUSINESS_RULE_ERROR
        - INTERNAL_ERROR
      description: Stable machine-readable code. Safe to branch on.
    ApiErrorDetailDto:
      type: object
      properties:
        code:
          description: Stable machine-readable code. Safe to branch on.
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Safe, user-facing message. Never contains a stack trace or internal detail.
      required:
        - code
        - message
    ApiErrorResponseDto:
      type: object
      properties:
        success:
          type: boolean
          example: false
          description: Always false on an error response.
        error:
          $ref: '#/components/schemas/ApiErrorDetailDto'
        requestId:
          type: string
          nullable: true
          description: Correlation id echoed from the x-request-id request header, or null.
      required:
        - success
        - error
        - requestId
...
    CountMetaDto:
      type: object
      properties:
        count:
          type: number
          description: Number of items in `data`.
      required:
        - count
...
    NotificationResponseDto:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        body:
          type: string
        data:
          type: object
          additionalProperties: true
          nullable: true
        isRead:
          type: boolean
        createdAt:
          format: date-time
          type: string
        readAt:
          format: date-time
          type: string
          nullable: true
      required:
        - id
        - type
        - title
        - body
        - data
        - isRead
        - createdAt
        - readAt
    NotificationMetaDto:
      type: object
      properties:
        unreadCount:
          type: number
          description: Total unread notifications for the user, regardless of filter.
      required:
        - unreadCount
    MarkedResultDto:
      type: object
      properties:
        marked:
          type: number
          description: How many notifications were marked as read by this call.
      required:
        - marked
    PricePointDto:
      type: object
      properties:
        lowest:
          type: number
        highest:
          type: number
        average:
          type: number
        capturedAt:
          type: string
      required:
        - lowest
        - highest
        - average
        - capturedAt
    PriceGraphPointDto:
      type: object
      properties:
        price:
          type: number
          description: Observed price for this seller at capture time.
        mrp:
          type: number
          nullable: true
          description: Seller-listed MRP at capture time, when the provider supplied one.
        discountPct:
          type: number
          nullable: true
          description: Discount off the seller-listed price at capture time. Not a drop relative to an earlier OmniSift observation.
        capturedAt:
          type: string
          format: date-time
      required:
        - price
        - mrp
        - discountPct
        - capturedAt
    PriceGraphSeriesDto:
      type: object
      properties:
        sellerId:
          type: string
        sellerName:
          type: string
        points:
          description: Chronological points for this seller.
          type: array
          items:
            $ref: '#/components/schemas/PriceGraphPointDto'
      required:
        - sellerId
        - sellerName
        - points
    PriceGraphResponseDto:
      type: object
      properties:
        series:
          description: One series per seller that has recorded history in the requested range.
          type: array
          items:
            $ref: '#/components/schemas/PriceGraphSeriesDto'
      required:
        - series
    PriceGraphMetaDto:
      type: object
      properties:
        sellerCount:
          type: number
          description: Number of seller series returned.
        days:
          type: number
          description: The range actually applied, after clamping.
      required:
        - sellerCount
        - days
...